Refactor tests: remove unused vars, fix assertion types
Cleaned up test code by removing unused variables in several test methods and replacing with discards where appropriate. Updated an assertion in EnumerableExtensionsTests to compare int values directly, reflecting the method's return type. These changes improve code clarity and maintainability.

+++ source/Unit Tests/dotNetTips.Spargine.Extensions.Tests/EnumerableExtensionsTests.cs
@@ -1210,7 +1210,7 @@ public class EnumerableExtensionsTests
 
 		var result = people.FastLongCount(p => p != null);
 
-		Assert.AreEqual((long)Count, result);
+		Assert.AreEqual(Count, result);
 	}
 
 	[TestMethod]
